Split integration tests in without and with auth.
