+++ .circleci/config.yml
@@ -27,11 +27,15 @@ jobs:
 
       - run:
           name: Build
-          command: cargo build --features without-auth
+          command: cargo build
 
       - run:
           name: Test
-          command: cargo test --features without-auth
+          command: cargo test
+
+      - run:
+          name: Test Integration
+          command: cargo test --test integration-no-auth -- --ignored
 
       - save_cache:
           key: rust-toolchain-{{ checksum "~/.rustup/settings.toml" }}
